Move report sources into their own package
