+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/deprecation/DefaultDeprecationBuilder.java
@@ -20,7 +20,7 @@ import org.gradle.api.problems.DocLink;
 import org.gradle.api.problems.deprecation.DeprecateMethodSpec;
 import org.gradle.api.problems.deprecation.DeprecatePluginSpec;
 import org.gradle.api.problems.deprecation.DeprecateSpec;
-import org.gradle.api.problems.deprecation.ReportSource;
+import org.gradle.api.problems.deprecation.source.ReportSource;
 import org.gradle.api.problems.internal.InternalProblem;
 import org.gradle.api.problems.internal.InternalProblemBuilder;
